beam update landsat 8

+++ beam/beam-landsat-reader/src/main/java/org/esa/beam/dataio/landsat/LandsatTMMetadata.java
@@ -36,7 +36,7 @@ public class LandsatTMMetadata {
     private List<MetadataElement> landsatMetadataElements;
 
     /**
-     * stores an integer value in a metadata attibute
+     * stores an integer value in a metadata attribute
      *
      * @param name
      * @param intData
@@ -49,7 +49,7 @@ public class LandsatTMMetadata {
     }
 
     /**
-     * stores a string value in a metadata attibute
+     * stores a string value in a metadata attribute
      *
      * @param name
      * @param stringData
@@ -62,7 +62,7 @@ public class LandsatTMMetadata {
     }
 
     /**
-     * stores an double value in a metadata attibute
+     * stores an double value in a metadata attribute
      *
      * @param name
      * @param doubleData
@@ -142,13 +142,13 @@ public class LandsatTMMetadata {
         if (data instanceof String) {
             attri = createAttribute(shortDescription, (String) data);
         } else if (data instanceof Integer) {
-            attri = createAttribute(shortDescription, ((Integer) data).intValue());
+            attri = createAttribute(shortDescription, ((Integer) data));
         } else if (data instanceof Double) {
-            attri = createAttribute(shortDescription, ((Double) data).doubleValue());
+            attri = createAttribute(shortDescription, (Double) data);
         } else if (data instanceof Long) {
-            attri = createAttribute(shortDescription, ((Long) data).longValue());
+            attri = createAttribute(shortDescription, ((Long) data));
         } else if (data instanceof Float) {
-            attri = createAttribute(shortDescription, ((Float) data).floatValue());
+            attri = createAttribute(shortDescription, ((Float) data));
         } else if (data instanceof ProductData.UTC) {
             attri = createAttribute(shortDescription, (ProductData.UTC) data);
         } else {
